+++ opt/jvmcommon.sh
@@ -6,25 +6,25 @@ calculate_java_memory_opts() {
   container_size=$(echo $CONTAINER_SIZE)
   case $container_size in
   S)
-    echo "$opts -Xmx160m -Xss512k -XX:CICompilerCount=2"
+    echo "$opts -Xmx160m -Xss512k -XX:CICompilerCount=1"
     ;;
   L)
     echo "$opts -Xmx671m -XX:CICompilerCount=2"
     ;;
   XL)
-    echo "$opts -Xmx1536m"
+    echo "$opts -Xmx1536m -XX:CICompilerCount=2"
     ;;
   2XL)
-    echo "$opts -Xmx3g"
+    echo "$opts -Xmx3g -XX:CICompilerCount=3"
     ;;
   3XL)
-    echo "$opts -Xmx6g"
+    echo "$opts -Xmx6g -XX:CICompilerCount=3"
     ;;
   4XL)
-    echo "$opts -Xmx12g"
+    echo "$opts -Xmx12g -XX:CICompilerCount=4"
     ;;
   *) # M Container and default if buildpack is used elsewhere
-    echo "$opts -Xmx300m -Xss512k -XX:CICompilerCount=2"
+    echo "$opts -Xmx300m -Xss512k -XX:CICompilerCount=1"
     ;;
   esac
 }
@@ -33,7 +33,8 @@ if [[ -d $HOME/.jdk ]]; then
   export JAVA_HOME="$HOME/.jdk"
   export PATH="$HOME/.scalingo/bin:$JAVA_HOME/bin:$PATH"
 else
-  export JAVA_HOME="$(realpath $(dirname $(which java))/..)"
+  JAVA_HOME="$(realpath "$(dirname "$(command -v java)")/..")"
+  export JAVA_HOME
 fi
 
 if [[ -d "$JAVA_HOME/jre/lib/amd64/server" ]]; then
@@ -42,7 +43,7 @@ elif [[ -d "$JAVA_HOME/lib/server" ]]; then
   export LD_LIBRARY_PATH="$JAVA_HOME/lib/server:$LD_LIBRARY_PATH"
 fi
 
-if cat "$JAVA_HOME/release" | grep -q '^JAVA_VERSION="1[0-9]'; then
+if grep -q '^JAVA_VERSION="1[0-9]' "$JAVA_HOME/release"; then
   default_java_mem_opts="$(calculate_java_memory_opts "-XX:+UseContainerSupport")"
 else
   default_java_mem_opts="$(calculate_java_memory_opts | sed 's/^ //')"